Enables inherited embargoes.
Treating all islandora objects of nodes in a tree, the RELS-EXT
information provides a means to traverse the tree. In the case of
inherited embargoes, it is only necessary to search upwards to find
(or not) a node with embargo data.

For a given object/node, its IP Embargo status will be one of:

- __Local embargo policy__: embargo is set on the object directly.
This overrides any embargo set on any ancestor node. This most
closely resembles the original functionality of this module.

- __Inherited embargo__: policy is inherited from nearest ancestor.

- __No embargo__: no applicable embargo policy exists, neither at
the local level nor at any ancestor level, all the way to the root.

Local embargo policy note: A node in an otherwise embargoed collection can be
whitelisted by choosing the 'empty' embargo list from the
drop-down, which previously has triggerd the deletion of the
embargo record from the db.